♻️ ScriptCodeDAO 去掉 extends Repo + save 容错 + OPFS mock 抽离
- ScriptCodeDAO 不再继承 Repo，过渡期间的 chrome.storage.local 操作
  改为 legacySave/legacyGet/legacyDelete 内部方法，过渡结束后直接删除
- save() 增加 try-catch，OPFS 写入失败不影响 chrome.storage.local
- OPFS mock 从 vitest.setup.ts 抽离到 tests/opfs_mock.ts
