+++ src/app/repo/scripts.ts
@@ -279,18 +279,15 @@ export class ScriptDAO extends Repo<Script> {
 
 // 为了防止脚本代码数据量过大,单独存储脚本代码
 // 内部使用 OPFS 优先存储，fallback 到 chrome.storage.local（过渡期间）
-export class ScriptCodeDAO extends Repo<ScriptCode> {
+export class ScriptCodeDAO {
+  private static readonly LEGACY_PREFIX = "scriptCode:";
   private _dirHandlePromise: Promise<FileSystemDirectoryHandle> | null = null;
-
-  constructor() {
-    super("scriptCode");
-  }
 
   private getDirHandle(): Promise<FileSystemDirectoryHandle> {
     if (!this._dirHandlePromise) {
       this._dirHandlePromise = navigator.storage
         .getDirectory()
-        .then((opfsRoot) => opfsRoot.getDirectoryHandle("storage_script_codes", { create: true }));
+        .then((opfsRoot) => opfsRoot.getDirectoryHandle("script_codes", { create: true }));
     }
     return this._dirHandlePromise;
   }
@@ -305,10 +302,15 @@ export class ScriptCodeDAO extends Repo<ScriptCode> {
   }
 
   public async save(val: ScriptCode): Promise<ScriptCode> {
-    // 写入 OPFS
-    await this.saveToOPFS(val);
+    // 写入 OPFS（失败不影响 chrome.storage.local）
+    try {
+      await this.saveToOPFS(val);
+    } catch {
+      // OPFS 写入失败，忽略
+    }
     // 过渡期间同步写入 chrome.storage.local
-    return super._save(val.uuid, val);
+    await this.legacySave(val);
+    return val;
   }
 
   public async get(key: string): Promise<ScriptCode | undefined> {
@@ -321,9 +323,9 @@ export class ScriptCodeDAO extends Repo<ScriptCode> {
     } catch {
       // OPFS 没有，fallback 到 chrome.storage.local
     }
-    const result = await super.get(key);
+    const result = await this.legacyGet(key);
     if (result) {
-      // 懒迁移：写入 OPFS（不 await，异步后台执行即可）
+      // 懒迁移：写入 OPFS
       this.saveToOPFS(result).catch(() => {});
     }
     return result;
@@ -342,7 +344,7 @@ export class ScriptCodeDAO extends Repo<ScriptCode> {
       // 忽略删除失败
     }
     // 过渡期间同步删除 chrome.storage.local
-    return super.delete(key);
+    await this.legacyDelete([key]);
   }
 
   public async deletes(keys: string[]): Promise<void> {
@@ -362,6 +364,62 @@ export class ScriptCodeDAO extends Repo<ScriptCode> {
       // 忽略
     }
     // 过渡期间同步删除 chrome.storage.local
-    return super.deletes(keys);
+    await this.legacyDelete(keys);
+  }
+
+  // --- 过渡期间 chrome.storage.local 操作，过渡结束后删除 ---
+
+  // 从 chrome.storage.local 读取所有脚本代码（仅迁移使用）
+  public all(): Promise<ScriptCode[]> {
+    return new Promise((resolve) => {
+      chrome.storage.local.get(null, (items) => {
+        if (chrome.runtime.lastError) {
+          console.error("chrome.storage.local.get error:", chrome.runtime.lastError);
+        }
+        const result: ScriptCode[] = [];
+        for (const key in items) {
+          if (key.startsWith(ScriptCodeDAO.LEGACY_PREFIX)) {
+            result.push(items[key]);
+          }
+        }
+        resolve(result);
+      });
+    });
+  }
+
+  private legacySave(val: ScriptCode): Promise<void> {
+    const key = ScriptCodeDAO.LEGACY_PREFIX + val.uuid;
+    return new Promise((resolve) => {
+      chrome.storage.local.set({ [key]: val }, () => {
+        if (chrome.runtime.lastError) {
+          console.error("chrome.storage.local.set error:", chrome.runtime.lastError);
+        }
+        resolve();
+      });
+    });
+  }
+
+  private legacyGet(key: string): Promise<ScriptCode | undefined> {
+    const storageKey = ScriptCodeDAO.LEGACY_PREFIX + key;
+    return new Promise((resolve) => {
+      chrome.storage.local.get(storageKey, (items) => {
+        if (chrome.runtime.lastError) {
+          console.error("chrome.storage.local.get error:", chrome.runtime.lastError);
+        }
+        resolve(items[storageKey]);
+      });
+    });
+  }
+
+  private legacyDelete(keys: string[]): Promise<void> {
+    const storageKeys = keys.map((key) => ScriptCodeDAO.LEGACY_PREFIX + key);
+    return new Promise((resolve) => {
+      chrome.storage.local.remove(storageKeys, () => {
+        if (chrome.runtime.lastError) {
+          console.error("chrome.storage.local.remove error:", chrome.runtime.lastError);
+        }
+        resolve();
+      });
+    });
   }
 }
